+++ rust/stake_neuron_from_cli/Cargo.toml
@@ -11,3 +11,8 @@ path = "src/main.rs"
 ic-agent = "0.35"
 tokio = { version = "1", features = ["full"] }
 clap = { version = "4", features = ["derive"] }
+candid = "0.10"
+sha2 = "0.10"
+hex = "0.4"
+serde = { version = "1.0", features = ["derive"] }
+crc32fast = "1.3"
